requirements.txt
lxml
requests
PyYAML

